Return data is optional for get_example_model, lots of restructuring because of circular imports.
