+++ development/documentation/scalability/run_single_scalability_exercise.py
@@ -3,6 +3,8 @@ import json
 import os
 import sys
 
+import respy.interface
+
 
 def main():
     """Evaluate the criterion function multiple times for a scalability report.
@@ -33,7 +35,7 @@ def main():
     import respy as rp
 
     # Get model
-    options, params, _ = rp.get_example_model(model)
+    options, params = respy.interface.get_example_model(model, with_data=False)
 
     # Simulate the data
     simulate = rp.get_simulate_func(params, options)
