+++ respy/tests/utils.py
@@ -1,12 +1,12 @@
 import numpy as np
 
-from respy.shared import get_example_model
+from respy import get_example_model
 from respy.tests.random_model import generate_random_model
 
 
 def process_model_or_seed(model_or_seed, **kwargs):
     if isinstance(model_or_seed, str):
-        params, options, _ = get_example_model(model_or_seed)
+        params, options = get_example_model(model_or_seed, with_data=False)
     else:
         np.random.seed(model_or_seed)
         params, options = generate_random_model(**kwargs)
